feat: requested for pokemon details
(pr feedback)
- created a separate files for info freezed model
- removes unneccessary white space
- rename model name to remove the prefixes
