+++ lib/api/model/ability_info.dart
@@ -1,17 +1,7 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-part 'pokemon_ability.freezed.dart';
-part 'pokemon_ability.g.dart';
-
-@freezed
-class PokemonAbility with _$PokemonAbility {
-  const factory PokemonAbility({
-    @JsonKey(name: 'ability') required AbilityInfo ability,
-  }) = _PokemonAbility;
-
-  factory PokemonAbility.fromJson(Map<String, dynamic> json) =>
-      _$PokemonAbilityFromJson(json);
-}
+part 'ability_info.freezed.dart';
+part 'ability_info.g.dart';
 
 @freezed
 class AbilityInfo with _$AbilityInfo {
@@ -19,7 +9,6 @@ class AbilityInfo with _$AbilityInfo {
     @JsonKey(name: 'name') required String name,
     @JsonKey(name: 'url') required String url,
   }) = _AbilityInfo;
-
   factory AbilityInfo.fromJson(Map<String, dynamic> json) =>
       _$AbilityInfoFromJson(json);
 }
