+++ lib/api/model/move_info.dart
@@ -1,17 +1,7 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-part 'pokemon_move.freezed.dart';
-part 'pokemon_move.g.dart';
-
-@freezed
-class PokemonMove with _$PokemonMove {
-  const factory PokemonMove({
-    @JsonKey(name: 'move') required MoveInfo move,
-  }) = _PokemonMove;
-
-  factory PokemonMove.fromJson(Map<String, dynamic> json) =>
-      _$PokemonMoveFromJson(json);
-}
+part 'move_info.freezed.dart';
+part 'move_info.g.dart';
 
 @freezed
 class MoveInfo with _$MoveInfo {
@@ -19,7 +9,6 @@ class MoveInfo with _$MoveInfo {
     @JsonKey(name: 'name') required String name,
     @JsonKey(name: 'url') required String url,
   }) = _MoveInfo;
-
   factory MoveInfo.fromJson(Map<String, dynamic> json) =>
       _$MoveInfoFromJson(json);
 }
