+++ lib/api/model/pokemon_details.dart
@@ -1,7 +1,7 @@
+import 'package:pokedex_async_redux/api/model/ability.dart';
+import 'package:pokedex_async_redux/api/model/move.dart';
+import 'package:pokedex_async_redux/api/model/stats.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:pokedex_async_redux/api/model/pokemon_ability.dart';
-import 'package:pokedex_async_redux/api/model/pokemon_move.dart';
-import 'package:pokedex_async_redux/api/model/pokemon_stats.dart';
 
 part 'pokemon_details.freezed.dart';
 part 'pokemon_details.g.dart';
@@ -11,11 +11,10 @@ class PokemonDetails with _$PokemonDetails {
   const factory PokemonDetails({
     @JsonKey(name: 'height') required double height,
     @JsonKey(name: 'weight') required double weight,
-    @JsonKey(name: 'abilities') required List<PokemonAbility> abilities,
-    @JsonKey(name: 'moves') required List<PokemonMove> moves,
-    @JsonKey(name: 'stats') required List<PokemonStats> stats,
+    @JsonKey(name: 'abilities') required List<Ability> abilities,
+    @JsonKey(name: 'moves') required List<Move> moves,
+    @JsonKey(name: 'stats') required List<Stats> stats,
   }) = _PokemonDetails;
-
   factory PokemonDetails.fromJson(Map<String, dynamic> json) =>
       _$PokemonDetailsFromJson(json);
 }
