--- lib/api/model/pokemon_stats.dart
@@ -1,27 +1,0 @@
-import 'package:freezed_annotation/freezed_annotation.dart';
-
-part 'pokemon_stats.freezed.dart';
-part 'pokemon_stats.g.dart';
-
-@freezed
-class PokemonStats with _$PokemonStats {
-  const factory PokemonStats({
-    @JsonKey(name: 'base_stat') required double baseStat,
-    @JsonKey(name: 'effort') required double effort,
-    @JsonKey(name: 'stat') required StatInfo stat,
-  }) = _PokemonStats;
-
-  factory PokemonStats.fromJson(Map<String, dynamic> json) =>
-      _$PokemonStatsFromJson(json);
-}
-
-@freezed
-class StatInfo with _$StatInfo {
-  const factory StatInfo({
-    @JsonKey(name: 'name') required String name,
-    @JsonKey(name: 'url') required String url,
-  }) = _StatInfo;
-
-  factory StatInfo.fromJson(Map<String, dynamic> json) =>
-      _$StatInfoFromJson(json);
-}
